Move ExampleRunner class under examples pacakge
