+++ examples/build.gradle.kts
@@ -7,6 +7,10 @@ plugins {
     id("com.diffplug.spotless") version "6.2.0"
 }
 
+group = "edu.cornell.cs.apl"
+
+val mainPackage = "${project.group}.${gradle.includedBuild("viaduct").name}.${project.name}"
+
 java {
     toolchain {
         languageVersion.set(JavaLanguageVersion.of(11))
@@ -29,7 +33,7 @@ dependencies {
 }
 
 application {
-    mainClass.set("edu.cornell.cs.apl.viaduct.codegeneration.CodegenRunnerKt")
+    mainClass.set("$mainPackage.ExampleRunnerKt")
 }
 
 tasks.withType<Test> {
@@ -49,6 +53,7 @@ spotless {
 
 val generateViaductProgramList by tasks.registering(GenerateViaductProgramList::class) {
     sourceDirectory.set(tasks.compileViaduct.map { it.sourceDirectory }.get())
+    outputPackage.set(mainPackage)
 }
 
 kotlin.sourceSets.main {
@@ -58,6 +63,9 @@ kotlin.sourceSets.main {
 abstract class GenerateViaductProgramList : DefaultTask() {
     @get:InputDirectory
     abstract val sourceDirectory: DirectoryProperty
+
+    @get:Input
+    abstract val outputPackage: Property<String>
 
     @OutputDirectory
     val outputDirectory: DirectoryProperty =
@@ -82,10 +90,13 @@ abstract class GenerateViaductProgramList : DefaultTask() {
             if (packageName.isEmpty()) className else "$packageName.$className"
         }.sorted()
 
-        outputDirectory.get().asFile.mkdirs()
-        val outputFile = outputDirectory.file("ViaductPrograms.kt").get().asFile
+
+        val packageDirectory = outputDirectory.get().asFile.resolve(outputPackage.get().replace(".", File.separator))
+        val outputFile = packageDirectory.resolve("ViaductPrograms.kt")
+        packageDirectory.mkdirs()
 
         val programsBlock = programs.map { "    $it" }.joinToString(",\n")
-        outputFile.writeText("package edu.cornell.cs.apl.viaduct.codegeneration\n\nval viaductPrograms = listOf(\n$programsBlock\n)\n")
+        val programsDeclaration = "val viaductPrograms = listOf(\n$programsBlock\n)\n"
+        outputFile.writeText("package ${outputPackage.get()}\n\n$programsDeclaration")
     }
 }
